rm: scan.go
